+++ CHANGELOG.md
@@ -3,7 +3,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.4.0 (12/25/17)
-- Add support for CustomField API end points.
+- Add support for CustomField, CustomRedirect API end points.
 
 ## 0.3.0 (12/15/17)
 - Add support for Visitor and VisitorActivity API endpoints.